fix(domain/authenticationConfiguration): token expiration dataType to double

+++ Domain/Models/AuthenticationConfiguration.cs
@@ -3,7 +3,7 @@
     public class AuthenticationConfiguration
     {
         public string AccessTokenSecret { get; set; }
-        public int AccessTokenExpirationMinutes { get; set; }
+        public double AccessTokenExpirationMinutes { get; set; }
         public string RefreshTokenSecret { get; set; }
         public double RefreshTokenExpirationMinutes { get; set; }
         public string Issuer { get; set; }
